backend/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
redis==5.0.8
celery==5.4.0
python-dotenv==1.0.1
openpyxl==3.1.5
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.22
bcrypt==4.0.1

numpy
opencv-python
scikit-image
scikit-video
scipy
